pyproject.toml
[project]
name = "ask-and-answer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.23.1",
    "msal>=1.33.0",
    "openai-agents>=0.2.3",
    "pymongo>=4.13.2",
    "langchain>=0.0.350",
    "langchain-community>=0.0.10",
    "unstructured[pdf]>=0.10.30",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
    "pdfminer.six>=20221105",
    "pi-heif>=1.0.0",
    "unstructured-inference>=1.0.5",
    "unstructured-pytesseract>=0.3.15",
    "gradio>=5.41.0",
    "langfuse>=3.2.2",
    "pydantic>=2.11.7",
]
